안우진: 경고 삭제

+++ CrazyArcade/GameEngineContents/BossLevel.cpp
@@ -59,7 +59,7 @@ void BossLevel::Loading()
 				}
 
 				//                          5 7
-				BlockSet->CreateBlock(float4(x * 40, y * 40), Name);
+				BlockSet->CreateBlock(float4(static_cast<float>(x * 40),static_cast<float>( y * 40)), Name);
 			}
 		}
 	}
