+++ CrazyArcade/GameEngineContents/Monster1Level.cpp
@@ -69,7 +69,7 @@ void Monster1Level::Loading()
 				}
 
 				//                          5 7
-				BlockSet->CreateBlock(float4(x * 40, y * 40), Name);
+				BlockSet->CreateBlock(float4(static_cast<float>(x * 40), static_cast<float>(y * 40)), Name);
 			}
 		}
 	}
